Add Electricity spot price widget
This widget supports showing electricity spot price for many areas in
Europe. However, there's no configuration UI yet. I'm hoping there will
be native support for editing the values in `prefs` later, but I'm also
open to implementing the configuration UI in the widget if I get some
help.

The chart would be better if displayed with timestamps, but that doesn't
currently work when the timestamps span two days.